cast nHeight to int32_t to match the VarIntMode expected type/value

See merge request bitcoinunlimited/BCHUnlimited!2705

+++ src/undo.h
@@ -27,7 +27,8 @@ public:
     template <typename Stream>
     void Serialize(Stream &s) const
     {
-        ::Serialize(s, VARINT(txout->nHeight * 2 + (txout->fCoinBase ? 1 : 0), VarIntMode::NONNEGATIVE_SIGNED));
+        ::Serialize(
+            s, VARINT(((int32_t)txout->nHeight) * 2 + (txout->fCoinBase ? 1 : 0), VarIntMode::NONNEGATIVE_SIGNED));
         if (txout->nHeight > 0)
         {
             // Required to maintain compatibility with older undo format.
